scripts/flickrDownload/requirements.txt
flickr_api
python-resize-image